refactor(sqlparser): move create source v1 compatible code to its own mod
